Pin to `setuptools<60.0.0`
NumPy does not support `setuptools >=60.0, because that is the
release that re-enabled the vendored `distutils` copy in
`setuptools`, and it has backwards compatibility issues
with `numpy.distutils`. There is currently no plan to resolve those;
NumPy itself will keep this pin, and other projects are advised to
migrate away from `numpy.distutils`.